requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
pyenchant==3.2.2
python-dotenv==0.21.1
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
rutracker-api==0.22.92
sniffio==1.3.0
soupsieve==2.4
urllib3==1.26.14
yarl==1.8.2
